elastic solve
