+++ src/constitutive/solid/elastic/mod.rs
@@ -18,6 +18,7 @@ mod almansi_hamel;
 pub use almansi_hamel::AlmansiHamel;
 
 use super::*;
+use crate::math::optimize::{NewtonRaphson, SecondOrder};
 
 /// Required methods for elastic constitutive models.
 pub trait Elastic<'a>
@@ -148,5 +149,74 @@ where
                 &deformation_gradient_inverse,
                 &second_piola_kirchoff_stress,
             ))
+    }
+    /// Solve for the unknown components of the Cauchy stress and deformation gradient under an applied load.
+    fn solve(
+        &self,
+        applied_load: AppliedLoad,
+    ) -> Result<(DeformationGradient, CauchyStress), ConstitutiveError> {
+        let optimization = NewtonRaphson {
+            ..Default::default()
+        };
+        let deformation_gradient = match applied_load {
+            AppliedLoad::UniaxialStress(deformation_gradient_11) => {
+                let deformation_gradient_33 = optimization.minimize(
+                    |deformation_gradient_33: &Scalar| {
+                        Ok(self.calculate_cauchy_stress(&DeformationGradient::new([
+                            [deformation_gradient_11, 0.0, 0.0],
+                            [0.0, *deformation_gradient_33, 0.0],
+                            [0.0, 0.0, *deformation_gradient_33],
+                        ]))?[2][2])
+                    },
+                    |deformation_gradient_33: &Scalar| {
+                        Ok(
+                            self.calculate_cauchy_tangent_stiffness(&DeformationGradient::new([
+                                [deformation_gradient_11, 0.0, 0.0],
+                                [0.0, *deformation_gradient_33, 0.0],
+                                [0.0, 0.0, *deformation_gradient_33],
+                            ]))?[2][2][2][2],
+                        )
+                    },
+                    1.0 / deformation_gradient_11.sqrt(),
+                    None,
+                    None,
+                )?;
+                DeformationGradient::new([
+                    [deformation_gradient_11, 0.0, 0.0],
+                    [0.0, deformation_gradient_33, 0.0],
+                    [0.0, 0.0, deformation_gradient_33],
+                ])
+            }
+            AppliedLoad::BiaxialStress(deformation_gradient_11, deformation_gradient_22) => {
+                let deformation_gradient_33 = optimization.minimize(
+                    |deformation_gradient_33: &Scalar| {
+                        Ok(self.calculate_cauchy_stress(&DeformationGradient::new([
+                            [deformation_gradient_11, 0.0, 0.0],
+                            [0.0, deformation_gradient_22, 0.0],
+                            [0.0, 0.0, *deformation_gradient_33],
+                        ]))?[2][2])
+                    },
+                    |deformation_gradient_33: &Scalar| {
+                        Ok(
+                            self.calculate_cauchy_tangent_stiffness(&DeformationGradient::new([
+                                [deformation_gradient_11, 0.0, 0.0],
+                                [0.0, deformation_gradient_22, 0.0],
+                                [0.0, 0.0, *deformation_gradient_33],
+                            ]))?[2][2][2][2],
+                        )
+                    },
+                    1.0 / deformation_gradient_11 / deformation_gradient_22,
+                    None,
+                    None,
+                )?;
+                DeformationGradient::new([
+                    [deformation_gradient_11, 0.0, 0.0],
+                    [0.0, deformation_gradient_22, 0.0],
+                    [0.0, 0.0, deformation_gradient_33],
+                ])
+            }
+        };
+        let cauchy_stress = self.calculate_cauchy_stress(&deformation_gradient)?;
+        Ok((deformation_gradient, cauchy_stress))
     }
 }
