+++ src/constitutive/solid/hyperelastic/mod.rs
@@ -34,16 +34,7 @@ pub use self::{
     neo_hookean::NeoHookean, saint_venant_kirchoff::SaintVenantKirchoff, yeoh::Yeoh,
 };
 use super::{elastic::Elastic, *};
-use crate::math::optimize::{NewtonRaphson, SecondOrder};
 use std::fmt::Debug;
-
-/// Possible applied loads.
-pub enum AppliedLoad {
-    /// Uniaxial stress given $`F_{11}`$.
-    UniaxialStress(Scalar),
-    /// Biaxial stress given $`F_{11}`$ and $`F_{22}`$.
-    BiaxialStress(Scalar, Scalar),
-}
 
 /// Required methods for hyperelastic constitutive models.
 pub trait Hyperelastic<'a>
@@ -59,73 +50,4 @@ where
         &self,
         deformation_gradient: &DeformationGradient,
     ) -> Result<Scalar, ConstitutiveError>;
-    /// Solve for the unknown components of the Cauchy stress and deformation gradient under an applied load.
-    fn solve(
-        &self,
-        applied_load: AppliedLoad,
-    ) -> Result<(DeformationGradient, CauchyStress), ConstitutiveError> {
-        let optimization = NewtonRaphson {
-            ..Default::default()
-        };
-        let deformation_gradient = match applied_load {
-            AppliedLoad::UniaxialStress(deformation_gradient_11) => {
-                let deformation_gradient_33 = optimization.minimize(
-                    |deformation_gradient_33: &Scalar| {
-                        Ok(self.calculate_cauchy_stress(&DeformationGradient::new([
-                            [deformation_gradient_11, 0.0, 0.0],
-                            [0.0, *deformation_gradient_33, 0.0],
-                            [0.0, 0.0, *deformation_gradient_33],
-                        ]))?[2][2])
-                    },
-                    |deformation_gradient_33: &Scalar| {
-                        Ok(
-                            self.calculate_cauchy_tangent_stiffness(&DeformationGradient::new([
-                                [deformation_gradient_11, 0.0, 0.0],
-                                [0.0, *deformation_gradient_33, 0.0],
-                                [0.0, 0.0, *deformation_gradient_33],
-                            ]))?[2][2][2][2],
-                        )
-                    },
-                    1.0 / deformation_gradient_11.sqrt(),
-                    None,
-                    None,
-                )?;
-                DeformationGradient::new([
-                    [deformation_gradient_11, 0.0, 0.0],
-                    [0.0, deformation_gradient_33, 0.0],
-                    [0.0, 0.0, deformation_gradient_33],
-                ])
-            }
-            AppliedLoad::BiaxialStress(deformation_gradient_11, deformation_gradient_22) => {
-                let deformation_gradient_33 = optimization.minimize(
-                    |deformation_gradient_33: &Scalar| {
-                        Ok(self.calculate_cauchy_stress(&DeformationGradient::new([
-                            [deformation_gradient_11, 0.0, 0.0],
-                            [0.0, deformation_gradient_22, 0.0],
-                            [0.0, 0.0, *deformation_gradient_33],
-                        ]))?[2][2])
-                    },
-                    |deformation_gradient_33: &Scalar| {
-                        Ok(
-                            self.calculate_cauchy_tangent_stiffness(&DeformationGradient::new([
-                                [deformation_gradient_11, 0.0, 0.0],
-                                [0.0, deformation_gradient_22, 0.0],
-                                [0.0, 0.0, *deformation_gradient_33],
-                            ]))?[2][2][2][2],
-                        )
-                    },
-                    1.0 / deformation_gradient_11 / deformation_gradient_22,
-                    None,
-                    None,
-                )?;
-                DeformationGradient::new([
-                    [deformation_gradient_11, 0.0, 0.0],
-                    [0.0, deformation_gradient_22, 0.0],
-                    [0.0, 0.0, deformation_gradient_33],
-                ])
-            }
-        };
-        let cauchy_stress = self.calculate_cauchy_stress(&deformation_gradient)?;
-        Ok((deformation_gradient, cauchy_stress))
-    }
 }
